+++ templates/mongo/config/environments/test.js
@@ -1,0 +1,5 @@
+app.configure('test', function(){
+    app.use(require('express').errorHandler({ dumpExceptions: true, showStack: true }));
+    app.settings.quiet = true;
+});
+