+++ templates/server_mongo.js
@@ -1,34 +1,10 @@
-/**
- * Module dependencies.
- */
-
-var express = require('express');
-
-var app = module.exports = express.createServer(),
-    mongoStore = require('connect-mongodb');
-
-app.settings.db = JSON.parse(require('fs').readFileSync(__dirname + '/config/database.json', 'utf-8'))[app.settings.env];
-
-var mongoSessionStore = mongoStore({
-    // maxAge:   60000,
-    dbname:   app.settings.db.database,
-    host:     app.settings.db.host,
-    username: app.settings.db.user,
-    password: app.settings.db.password
-}, function () {});
-
-app.configure(function(){
-    app.use(express.static(__dirname + '/public'));
-    app.set('views', __dirname + '/app/views');
-    app.set('view engine', 'ejs');
-    app.use(express.bodyParser());
-    app.use(express.cookieParser());
-    app.use(express.session({secret: 'secret', store: mongoSessionStore}));
-    app.use(express.methodOverride());
-    app.use(app.router);
-});
+var express = require('express'),
+    app = module.exports = express.createServer();
 
 app.configure('development', function(){
+    app.disable('view cache');
+    app.disable('model cache');
+    app.disable('eval cache');
     app.use(express.errorHandler({ dumpExceptions: true, showStack: true }));
 });
 
@@ -42,6 +18,9 @@ app.configure('test', function(){
 });
 
 app.configure('production', function(){
+    app.enable('view cache');
+    app.enable('model cache');
+    app.enable('eval cache');
     app.use(express.errorHandler());
     app.settings.quiet = true;
 });